Module paralegal_policy::algo::ahb

Checking always-happens-before relationships

Structs

ALWAYS_HAPPENS_BEFORE_NAME 🔒

AlwaysHappensBefore

Statistics about the result of running crate::Context::always_happens_before that are useful to understand how the property failed.

Tracer 🔒

Enums

Trace 🔒

TraceLevel

Retention level of additional information about the execution of an always_happens_before.

Type Aliases

DefId

LocalDefId