paralegal_policy::algo
Module flows_to

Precomputed reachability queries

Structs

CtrlFlowsTo

Precomputed indices for common queries of a PDG.

DataAndControlInfluencees

An Iterator over the SPDGNodes from the given src in the transitive closure of data and control flow of the given SPDG.