paralegal_policy
Module context

Modules

private 🔒

Structs

DisplayDef

Provide display trait for DefId in a Context.

MarkerTargets

Collection of entities a particular marker has been applied to

RootContext

Interface for defining policies.

Traits

Context

Actions that behave differently depending on the context

NodeExt

Extension trait with queries for single nodes

NodeQueries

Context queries conveniently accessible on nodes

Type Aliases

DefId

FlowsTo 🔒

FunctionId

The type identifying a function that is used in call sites.

LocalDefId

MarkableId

Identifier for a graph element that allows attaching a marker.

Marker

User-defined PDG markers.

MarkerIndex 🔒