Trait paralegal_policy::context::NodeExt

pub trait NodeExt: Sealed {
    // Required methods
    fn has_type(self, t: TypeId, ctx: &Context) -> bool;
    fn associated_call_site(self, ctx: &Context) -> CallString;
    fn types(self, ctx: &Context) -> &[TypeId];
    fn describe(self, ctx: &Context) -> DisplayNode<'_>;
    fn info(self, ctx: &Context) -> &NodeInfo;
    fn instruction(self, ctx: &Context) -> &InstructionInfo;
    fn successors(
        self,
        ctx: &Context
    ) -> Box<dyn Iterator<Item = GlobalNode> + '_>;
    fn predecessors(
        self,
        ctx: &Context
    ) -> Box<dyn Iterator<Item = GlobalNode> + '_>;
    fn get_location(self, ctx: &Context) -> &Span;
    fn has_marker<C: HasDiagnosticsBase>(self, ctx: C, marker: Marker) -> bool;
    fn shortest_path(
        self,
        to: GlobalNode,
        ctx: &Context,
        edge_selection: EdgeSelection
    ) -> Option<Box<[GlobalNode]>>;
}

Extension trait with queries for single nodes

Required Methods

fn has_type(self, t: TypeId, ctx: &Context) -> bool

Returns true if this node has the provided type

fn associated_call_site(self, ctx: &Context) -> CallString

Find the call string for the statement or function that produced this node.

fn types(self, ctx: &Context) -> &[TypeId]

Get the type(s) of a Node.

fn describe(self, ctx: &Context) -> DisplayNode<'_>

Returns a DisplayNode for the given Node

fn info(self, ctx: &Context) -> &NodeInfo

Retrieve metadata about a node.

fn instruction(self, ctx: &Context) -> &InstructionInfo

Retrieve metadata about the instruction executed by a specific node.

fn successors(self, ctx: &Context) -> Box<dyn Iterator<Item = GlobalNode> + '_>

Return the immediate successors of this node

fn predecessors( self, ctx: &Context ) -> Box<dyn Iterator<Item = GlobalNode> + '_>

Return the immediate predecessors of this node

fn get_location(self, ctx: &Context) -> &Span

Get the span of a node

fn has_marker<C: HasDiagnosticsBase>(self, ctx: C, marker: Marker) -> bool

Returns whether this Node has the marker applied to it directly or via its type.

fn shortest_path( self, to: GlobalNode, ctx: &Context, edge_selection: EdgeSelection ) -> Option<Box<[GlobalNode]>>

The shortest path between this and a target node