Struct paralegal_policy::diagnostics::ControllerContext

pub struct ControllerContext {
    id: Endpoint,
    inner: Arc<dyn HasDiagnosticsBase>,
}

A context for controllers

You may call any method and access any field defined on Context. In addition all diagnostics messages emitted from this struct will carry the name of the controller.

See the module level documentation for more information on diagnostic context management.

Fields
id: Endpointinner: Arc<dyn HasDiagnosticsBase>
Implementations

impl ControllerContext

pub fn named_combinator<A>( self: Arc<Self>, name: impl Into<Identifier>, computation: impl FnOnce(Arc<CombinatorContext>) -> A ) -> A

Add a named combinator to the diagnostic context.

See the module level documentation for more information on diagnostic context management.

pub fn named_policy<A>( self: Arc<Self>, name: impl Into<Identifier>, policy: impl FnOnce(Arc<PolicyContext>) -> A ) -> A

Add a policy to the diagnostic context.

See the module level documentation for more information on diagnostic context management.

pub fn id(&self) -> Endpoint

Access the id for the controller of this context

pub fn current(&self) -> &SPDG

Access the current controller contents

fn for_all(ctx: Arc<dyn HasDiagnosticsBase>) -> impl Iterator<Item = Arc<Self>>