paralegal_policy
Macro loc

macro_rules! loc {
    () => { ... };
    ($($t:tt)+) => { ... };
}

A convenience macro that uses file!, line! and column! to return the string "file:line:column". This can be used to mention policy source locations in policies.

If additional arguments are procided these are concat!ed to the end with a space in betwee the location and the rest.