paralegal_policy
Struct GlobalNode

pub struct GlobalNode {
    pub(crate) node: NodeIndex,
    pub(crate) controller_id: DefId,
}

A globally identified node in an SPDG

Fields
node: NodeIndexcontroller_id: DefId
Implementations

impl GlobalNode

pub fn unsafe_new(ctrl_id: DefId, index: usize) -> GlobalNode

Create a new node with no guarantee that it exists in the SPDG of the controller.

pub fn from_local_node(ctrl_id: DefId, node: NodeIndex) -> GlobalNode

Create a new globally identified node by pairing a node local to a particular SPDG with itโs controller id.

Meant for internal use only.

pub fn local_node(self) -> NodeIndex

The local node in the SPDG

pub fn controller_id(self) -> DefId

The identifier for the SPDG this node is contained in